DayNight lighting fixes && MESON/THERMAL CHANGES
LISTEN UP!!
In order for daynight lighting to work with mesons and thermals they needed to be changed slightly
instead of giving them full nightvision (which is the alternative)
I gave them a slight tint (green or red based on which) and a little bit of brightness. this prevents users using them as an alternative to night vision since its not full nightvision. If needed we can discuss this and change if needed.